requirements.txt
aiosignal==1.3.2
alembic==1.15.2
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
category_encoders==2.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dash==3.0.2
decorator==5.2.1
deprecation==2.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
greenlet==3.1.1
hyperopt==0.2.7
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipython==8.35.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kaleido==0.2.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
msgpack==1.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
optuna==4.2.1
optuna-integration==4.2.1
orjson==3.10.16
packaging==24.2
pandas==2.1.4
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt_toolkit==3.0.50
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyaml==25.1.0
pyarrow==19.0.1
pycaret==3.3.2
Pygments==2.19.1
pyod==2.0.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.44.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-optimize==0.10.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
tbats==1.1.3
tenacity==9.1.2
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
tune-sklearn==0.5.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wurlitzer==3.1.1
xxhash==3.5.0
yellowbrick==1.5
zipp==3.21.0
